sigma_rizzler.c: Make the enemy only heal when its health is lower than 60% of its max health.
With the current random behavior, the enemy sometimes always chooses to heal itself, despite its health still pretty high. Which makes the behavior a bit weird.

This commit adds a condition on top of the existing random one that will only let the enemy heal when its health is lower than a set value (eg. less than 60%). We can change this value depending on the difficulty. For example, make the enemy heal earlier on harder difficulty.